Add fused types to non-parallel version

+++ examples/test.py
@@ -13,7 +13,8 @@ im2 = relabel_sequential(im2.sum(axis=-1))[0]
 shape = (int(np.max(im1) + 1), int(np.max(im2) + 1))
 out = overlap_parallel(im1.astype(np.int32), im2.astype(np.int32), shape)
 print(out.sum())
-out_serial = overlap(im1.astype(np.int32), im2.astype(np.int32), shape)
+# out_serial = overlap(im1.astype(np.int32), im2.astype(np.int32), shape)
+out_serial = overlap(im1.astype(np.uint16), im2.astype(np.uint16), shape)
 
 # from IPython import embed
 # embed(colors="Linux")
